requirements.txt
flask==2.2.5
gunicorn==20.1.0
pandas==2.3.1
google-cloud-bigquery==3.35.0
db-dtypes==1.4.3